rest_api/requirements.txt
librosa
Django==3.2.12
djangorestframework==3.13.1
gunicorn==20.1.0
django-cors-headers==3.10.0
djangorestframework-simplejwt
dj-rest-auth
django-allauth
psycopg2
yt-dlp